Add: DetailFragment 추가
SessionFragment를 SessionSelect/DetailFragment로 다시 한번 구분
- NavigationDrawer에 Session Highlight를 유지하기 위함

Session에서 Info를 전달
- Parcelize
