+++ app/src/androidTest/java/com/example/ifkakao/presentation/session/SessionSelectViewModelTest.kt
@@ -15,7 +15,7 @@ import org.junit.Test
 import retrofit2.Retrofit
 import retrofit2.create
 
-class SessionViewModelTest {
+class SessionSelectViewModelTest {
     @OptIn(ExperimentalSerializationApi::class)
     private val sessionService: SessionService =
         Retrofit.Builder()
@@ -25,7 +25,7 @@ class SessionViewModelTest {
             .create()
     private val sessionRepositoryImpl = SessionRepositoryImpl(sessionService)
     private val sessionsUseCase = GetSessionsUseCase(sessionRepositoryImpl)
-    private val viewModel = SessionViewModel(sessionsUseCase)
+    private val viewModel = SessionSelectViewModel(sessionsUseCase)
 
     @Test
     fun testFilterInfoList() = runBlocking {
