+++ app/src/main/java/com/example/ifkakao/presentation/session/detail/DetailViewModel.kt
@@ -1,0 +1,8 @@
+package com.example.ifkakao.presentation.session.detail
+
+import androidx.lifecycle.ViewModel
+import com.example.ifkakao.domain.model.Info
+
+class DetailViewModel : ViewModel() {
+    var info: Info? = null
+}
