+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -18,5 +18,5 @@
     <fragment
         android:id="@+id/nav_session"
         android:name="com.example.ifkakao.presentation.session.SessionFragment"
-        tools:layout="@layout/fragment_session" />
+        tools:layout="@layout/fragment_session"/>
 </navigation>
